deploy: add DB-only option

+++ deploy.sh
@@ -3,9 +3,13 @@
 PARAMS='-i stop-bannon.pem'
 HOST='ubuntu@ec2-52-15-139-29.us-east-2.compute.amazonaws.com'
 
-ssh $PARAMS $HOST rm -rf website
-scp $PARAMS -Cr website $HOST:
-ssh $PARAMS $HOST cp local_settings.py website/website
+if [ -z $DB ]; then
+    ssh $PARAMS $HOST rm -rf website
+    scp $PARAMS -Cr website $HOST:
+    ssh $PARAMS $HOST cp local_settings.py website/website
+else
+    scp $PARAMS -C website/db.sqlite3 $HOST:website
+fi
 
 echo "\n\n"
